+++ src/main/java/kr/co/conceptbe/idea/domain/Idea.java
@@ -18,7 +18,7 @@ import kr.co.conceptbe.comment.Comment;
 import kr.co.conceptbe.common.entity.base.BaseTimeEntity;
 import kr.co.conceptbe.branch.domain.Branch;
 import kr.co.conceptbe.purpose.domain.Purpose;
-import kr.co.conceptbe.common.entity.domain.TeamRecruitment;
+import kr.co.conceptbe.teamrecruitment.domain.TeamRecruitment;
 import kr.co.conceptbe.idea.domain.vo.IdeaBranches;
 import kr.co.conceptbe.idea.domain.vo.IdeaPurposes;
 import kr.co.conceptbe.idea.domain.vo.IdeaTeamRecruitments;
