+++ src/main/java/kr/co/conceptbe/idea/domain/IdeaTeamRecruitment.java
@@ -8,7 +8,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kr.co.conceptbe.common.entity.domain.TeamRecruitment;
+import kr.co.conceptbe.teamrecruitment.domain.TeamRecruitment;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
